Windows: Tolerate MSVC mangled name in gdb.rocm/mi-aspace.exp
gdb.rocm/mi-aspace.exp shows one FAIL on Windows:

 -data-evaluate-expression &global_var
 ^done,value="0x7ff62b170c38 <?global_var@@3HA>"
 (gdb)
 FAIL: gdb.rocm/mi-aspace.exp: test_mi_memory_changed: get global_var address

The rest of the testcase is skipped due to that FAIL.

"?global_var@@3HA" is simply "global_var" in MSVC's mangling scheme,
which GDB does not yet understand.  Adjust the testcase to tolerate it
for now, so that the rest of tests can proceed.

Note this is written in a way that once GDB supports the MSVC mangling
scheme, we should get a KPASS, which will remind us to come back here
and clean up the testcase.

Change-Id: I8b92dbc796eca47a8decfc48dac862591b621004
commit-id:7b8b398e