gdb.rocm: explicitly set GPU coredump pattern
With the recent changes in rocr-runtime [1], GPU coredumps are
now generated with the same pattern as host's coredump with the
".gpu" suffix appended to it.  That's usually "core.<pid>.gpu".
This is next to the host coredump with "core.<pid>" pattern.
This does not fly well with "gdb.exp:core_find {}" procedure,
because when globbing for the coredumps with "core.*" pattern,
it expects only one entry.  As a result, it ignores the listing
and acts like no coredump is found.  That leads to skipping
tests like runtime-core.exp:

  WARNING: can't generate a core file - core tests
           suppressed - check ulimit -c

The same commit [1] produced a mechanism to enforce the desired
pattern for GPU coredumps through the "HSA_COREDUMP_PATTERN"
environment variable.  This patch sets that variable to the
previously working pattern: "gpucore.<pid>".  This step is added
to "rocm_core_find_1 {}" in "lib/rocm.exp", just before invoking
"lib/gdb.exp:core_find {}".

[1]
Commit title "rocr: GPU core file location support".